backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
pydantic
pydantic-settings
requests
